resolved github PR comments 2
